Refactor: Separate DataField render and handling logic
